requirements.txt
langchain-groq
langchain-community
langchain-core
langgraph
langgraph-checkpoint-postgres
tavily-python
python-dotenv
psycopg[binary]
flask
flask-cors
bcrypt
jwt
gunicorn